Update Get-WannaCryPatchState.ps1
The script has been modified so that it won't check computers that have already returned as patched. Computers returned as patched are appended to a separate file which the script will check on subsequent runs. 
Log file output has been changed to include the date as part of the file name.